# various performances improvements (30-50% in certain workflows, 15-25% in average)