@erocarrera 2022-03-06
Thanks! I had to fix a small issue in [AddressSet](https://github.com/erocarrera/pefile/commit/2a6e328ca33ef96ead3d04d8ec22648b90e80546) but it's a great PR!